+++ tests/unit/test_core/test_auth.py
@@ -1,8 +1,11 @@
 """Test kytos.core.auth module."""
 import base64
-import hashlib
 from unittest import TestCase
-from unittest.mock import Mock, patch
+from unittest.mock import MagicMock, Mock, patch
+
+from pydantic import ValidationError
+from pymongo.errors import DuplicateKeyError
+from werkzeug.exceptions import NotFound
 
 from kytos.core import Controller
 from kytos.core.auth import Auth
@@ -11,7 +14,6 @@ from kytos.core.config import KytosConfig
 
 KYTOS_CORE_API = "http://127.0.0.1:8181/api/kytos/"
 API_URI = KYTOS_CORE_API+"core"
-STOREHOUSE_API_URI = KYTOS_CORE_API+"storehouse/v1/kytos.core.auth.users"
 
 
 # pylint: disable=unused-argument
@@ -20,9 +22,11 @@ class TestAuth(TestCase):
 
     def setUp(self):
         """Instantiate a controller and an Auth."""
+        Auth.get_user_controller = MagicMock()
         self.patched_events = []  # {'event_name': box_object}
         self.server_name_url = 'http://localhost:8181/api/kytos'
         self.controller = self._get_controller_mock()
+        self.controller.start_auth()
         self.auth = Auth(self.controller)
         self.username, self.password = self._create_super_user()
         self.token = self._get_token()
@@ -33,12 +37,6 @@ class TestAuth(TestCase):
         }
         self.auth_header = {"Authorization": f"Bearer {self.token}"}
 
-    def _patch_event_trigger(self, event):
-        """Patch event callback trigger."""
-        for patched_event in self.patched_events:
-            box = patched_event.get(event.content.get('callback').__name__)
-            event.content.get('callback')(None, box, None)
-
     def _get_controller_mock(self):
         """Return a controller mock."""
         options = KytosConfig().options['daemon']
@@ -46,10 +44,6 @@ class TestAuth(TestCase):
         controller = Controller(options)
         controller.buffers = KytosBuffers()
         controller.log = Mock()
-
-        # Patch event callback trigger.
-        controller.buffers.app.put = self._patch_event_trigger
-
         return controller
 
     @staticmethod
@@ -57,24 +51,16 @@ class TestAuth(TestCase):
         """Return a flask api test client."""
         return auth.controller.api_server.app.test_client()
 
-    @patch('kytos.core.auth.Auth._create_superuser')
-    def _create_super_user(self, mock_username=None):
+    def _create_super_user(self):
         """Create a superuser to integration test."""
         username = "test"
         password = "password"
-        email = "test@kytos.io"
-
-        mock_username.return_value.get_username.return_value = username
-        mock_username.return_value.get_email.return_value = email
-        self.auth._create_superuser()  # pylint: disable=protected-access
-
         return username, password
 
     @patch('kytos.core.auth.Auth.get_jwt_secret', return_value="abc")
     def _get_token(self, mock_jwt_secret=None):
         """Make a request to get a token to be used in tests."""
-        box = Mock()
-        box.data = {
+        box = {
             # "password" digested
             'password': 'b109f3bbbc244eb82441917ed06d618b9008dd09b3befd1b5e073'
                         '94c706a8bb980b1d7785e5976ec049b46df5f1326af5a2ea6d103'
@@ -86,8 +72,7 @@ class TestAuth(TestCase):
                 bytes(self.username + ":" + self.password, "ascii")
             ).decode("ascii")
         }
-        # Patch _find_user_callback event callback.
-        self.patched_events.append({'_find_user_callback': box})
+        self.auth.user_controller.get_user.return_value = box
         url = f"{API_URI}/auth/login/"
         api = self.get_auth_test_client(self.auth)
         success_response = api.open(url, method='GET', headers=header)
@@ -119,20 +104,10 @@ class TestAuth(TestCase):
                 bytes("nonexistent" + ":" + "nonexistent", "ascii")
             ).decode("ascii")
         }
-        box = Mock()
-        box.data = {
-            # "password" digested
-            'password': 'b109f3bbbc244eb82441917ed06d618b9008dd09b3befd1b5e073'
-                        '94c706a8bb980b1d7785e5976ec049b46df5f1326af5a2ea6d103'
-                        'fd07c95385ffab0cacbc86'
-        }
-        # Patch _find_user_callback event callback.
-        self.patched_events.append({'_find_user_callback': box})
         url = f"{API_URI}/auth/login/"
         api = self.get_auth_test_client(self.auth)
         success_response = api.open(url, method='GET', headers=valid_header)
         error_response = api.open(url, method='GET', headers=invalid_header)
-
         self.assertEqual(success_response.status_code, 200)
         self.assertEqual(error_response.status_code, 401)
 
@@ -141,13 +116,11 @@ class TestAuth(TestCase):
         """Test auth list users endpoint."""
         invalid_header = {"Authorization": "Bearer invalidtoken"}
         schema = {"users": list}
-        password = "password".encode()
-        # Patch _list_users_callback event callback.
-        event_boxes = [self.user_data,
-                       {"username": "authtempuser2",
-                        "email": "tempuser2@kytos.io",
-                        "password": hashlib.sha512(password).hexdigest()}]
-        self.patched_events.append({'_list_users_callback': event_boxes})
+        response = {'users': [
+                        self.user_data['username'],
+                        {"username": "authtempuser2"}
+                    ]}
+        self.auth.user_controller.get_users.return_value = response
         api = self.get_auth_test_client(self.auth)
         url = f"{API_URI}/auth/users/"
         success_response = api.open(url, method='GET',
@@ -162,34 +135,40 @@ class TestAuth(TestCase):
     @patch('kytos.core.auth.Auth.get_jwt_secret', return_value="abc")
     def test_03_create_user_request(self, mock_jwt_secret):
         """Test auth create user endpoint."""
-        # Patch _create_user_callback event callback.
-        self.patched_events.append({'_create_user_callback': self.user_data})
         api = self.get_auth_test_client(self.auth)
         url = f"{API_URI}/auth/users/"
         success_response = api.open(url, method='POST', json=self.user_data,
                                     headers=self.auth_header)
-
-        self.assertEqual(success_response.status_code, 200)
+        self.assertEqual(success_response.status_code, 201)
 
     @patch('kytos.core.auth.Auth.get_jwt_secret', return_value="abc")
-    def test_03_create_user_request_error(self, mock_jwt_secret):
+    def test_03_create_user_request_conflict(self, mock_jwt_secret):
         """Test auth create user endpoint."""
-        # Patch _create_user_callback event callback.
-        self.patched_events.append({'_create_user_callback': None})
+        controller = self.auth.user_controller
+        controller.create_user.side_effect = DuplicateKeyError(0)
         api = self.get_auth_test_client(self.auth)
         url = f"{API_URI}/auth/users/"
         error_response = api.open(url, method='POST', json=self.user_data,
                                   headers=self.auth_header)
-
         self.assertEqual(error_response.status_code, 409)
+
+    @patch('kytos.core.auth.Auth.get_jwt_secret', return_value="abc")
+    def test_03_create_user_request_bad(self, mock_jwt_secret):
+        """Test auth create user endpoint."""
+        controller = self.auth.user_controller
+        controller.create_user.side_effect = ValidationError('', '')
+        api = self.get_auth_test_client(self.auth)
+        url = f"{API_URI}/auth/users/"
+        error_response = api.open(url, method='POST', json=self.user_data,
+                                  headers=self.auth_header)
+        self.assertEqual(error_response.status_code, 400)
 
     @patch('kytos.core.auth.Auth.get_jwt_secret', return_value="abc")
     def test_04_list_user_request(self, mock_jwt_secret):
         """Test auth list user endpoint."""
-        schema = {"data": {"email": str, "username": str}}
-        box = Mock()
-        box.data = self.user_data
-        self.patched_events.append({'_find_user_callback': box})
+        schema = {"email": str, "username": str}
+        box = self.user_data
+        self.auth.user_controller.get_user_nopw.return_value = box
         api = self.get_auth_test_client(self.auth)
         url = f"{API_URI}/auth/users/{self.user_data.get('username')}"
         success_response = api.open(url, method='GET',
@@ -202,18 +181,16 @@ class TestAuth(TestCase):
     @patch('kytos.core.auth.Auth.get_jwt_secret', return_value="abc")
     def test_04_list_user_request_error(self, mock_jwt_secret):
         """Test auth list user endpoint."""
-        self.patched_events.append({'_find_user_callback': None})
+        self.auth.user_controller.get_user_nopw.return_value = None
         api = self.get_auth_test_client(self.auth)
         url = f"{API_URI}/auth/users/user3"
         error_response = api.open(url, method='GET', headers=self.auth_header)
-
         self.assertEqual(error_response.status_code, 404)
 
     @patch('kytos.core.auth.Auth.get_jwt_secret', return_value="abc")
     def test_05_update_user_request(self, mock_jwt_secret):
         """Test auth update user endpoint."""
         data = {"email": "newemail_tempuser@kytos.io"}
-        self.patched_events.append({'_update_user_callback': data})
         api = self.get_auth_test_client(self.auth)
         url = f"{API_URI}/auth/users/{self.user_data.get('username')}"
         success_response = api.open(url, method='PATCH', json=data,
@@ -222,9 +199,9 @@ class TestAuth(TestCase):
         self.assertEqual(success_response.status_code, 200)
 
     @patch('kytos.core.auth.Auth.get_jwt_secret', return_value="abc")
-    def test_05_update_user_request_error(self, mock_jwt_secret):
+    def test_05_update_user_request_not_found(self, mock_jwt_secret):
         """Test auth update user endpoint."""
-        self.patched_events.append({'_update_user_callback': None})
+        self.auth.user_controller.update_user.return_value = {}
         api = self.get_auth_test_client(self.auth)
         url = f"{API_URI}/auth/users/user5"
         error_response = api.open(url, method='PATCH', json={},
@@ -233,10 +210,32 @@ class TestAuth(TestCase):
         self.assertEqual(error_response.status_code, 404)
 
     @patch('kytos.core.auth.Auth.get_jwt_secret', return_value="abc")
+    def test_05_update_user_request_bad(self, mock_jwt_secret):
+        """Test auth update user endpoint"""
+        controller = self.auth.user_controller
+        controller.update_user.side_effect = ValidationError('', '')
+        api = self.get_auth_test_client(self.auth)
+        url = f"{API_URI}/auth/users/user5"
+        error_response = api.open(url, method='PATCH', json={},
+                                  headers=self.auth_header)
+
+        self.assertEqual(error_response.status_code, 400)
+
+    @patch('kytos.core.auth.Auth.get_jwt_secret', return_value="abc")
+    def test_05_update_user_request_conflict(self, mock_jwt_secret):
+        """Test auth update user endpoint"""
+        controller = self.auth.user_controller
+        controller.update_user.side_effect = DuplicateKeyError(0)
+        api = self.get_auth_test_client(self.auth)
+        url = f"{API_URI}/auth/users/user5"
+        error_response = api.open(url, method='PATCH', json={},
+                                  headers=self.auth_header)
+
+        self.assertEqual(error_response.status_code, 409)
+
+    @patch('kytos.core.auth.Auth.get_jwt_secret', return_value="abc")
     def test_06_delete_user_request(self, mock_jwt_secret):
         """Test auth delete user endpoint."""
-        # Patch _delete_user_callback event callback.
-        self.patched_events.append({'_delete_user_callback': self.user_data})
         api = self.get_auth_test_client(self.auth)
         url = f"{API_URI}/auth/users/{self.user_data.get('username')}"
         success_response = api.open(url, method='DELETE',
@@ -247,11 +246,17 @@ class TestAuth(TestCase):
     @patch('kytos.core.auth.Auth.get_jwt_secret', return_value="abc")
     def test_06_delete_user_request_error(self, mock_jwt_secret):
         """Test auth delete user endpoint."""
-        # Patch _delete_user_callback event callback.
-        self.patched_events.append({'_delete_user_callback': None})
+        self.auth.user_controller.delete_user.return_value = {}
         api = self.get_auth_test_client(self.auth)
         url = f"{API_URI}/auth/users/nonexistent"
         success_response = api.open(url, method='DELETE',
                                     headers=self.auth_header)
 
         self.assertEqual(success_response.status_code, 404)
+
+    def test_07_find_user_error(self):
+        """Test _finb_user NotFound"""
+        self.auth.user_controller.get_user.return_value = None
+        with self.assertRaises(NotFound):
+            # pylint: disable=protected-access
+            self.auth._find_user("name")